+++ json/json_pointer.h
@@ -126,7 +126,7 @@ namespace cppdatalib
                                     return NULL;
                             }
 
-                            reference = &reference->get_array_unchecked().data()[array_index];
+                            reference = &reference->get_array_unchecked().data()[static_cast<size_t>(array_index)];
                         }
                         else
                         {
@@ -257,12 +257,12 @@ namespace cppdatalib
 
                             if (destroy_element && idx > pointer.size())
                             {
-                                reference->erase_element(array_index);
+                                reference->erase_element(static_cast<size_t>(array_index));
                                 return reference;
                             }
 
                             parent = reference;
-                            reference = &reference->get_array_ref().data()[array_index];
+                            reference = &reference->get_array_ref().data()[static_cast<size_t>(array_index)];
                         }
                         else
                         {
